Relocated lib
