+++ src/pages/view/[project].js
@@ -2,7 +2,7 @@ import React from 'react';
 import { useRouter } from 'next/router';
 import Head from 'next/head';
 import { getDataFromTree } from '@apollo/react-ssr';
-import withApollo from '../../lib/withApollo';
+import withApollo from '../../providers/withApollo';
 
 import View from '../../components/View';
